+++ files/default/datadog_php_ping.py
@@ -1,9 +1,12 @@
+# stdlib
 import re
 import urllib2
-import base64
+import urlparse
 
+# project
 from util import headers
 from checks import AgentCheck
+from checks.utils import add_basic_auth
 
 class PhpPing(AgentCheck):
     """Monitors php-fpm status via ping-url
@@ -13,46 +16,55 @@ class PhpPing(AgentCheck):
 
     """
 
-    def __init__(self, name, init_config, agentConfig, instances=None):
-        AgentCheck.__init__(self, name, init_config, agentConfig, instances)
-        self.assumed_url = {}
-
     def check(self, instance):
         if 'php_ping_url' not in instance:
             raise Exception('php instance missing "php_ping_url" value.')
-
-        url = self.assumed_url.get(instance['php_ping_url'], instance['php_ping_url'])
-
         tags = instance.get('tags', [])
-
-        req = urllib2.Request(url, None,
-            headers(self.agentConfig))
-        if 'php_ping_user' in instance and 'php_ping_password' in instance:
-            auth_str = '%s:%s' % (instance['php_ping_user'], instance['php_ping_password'])
-            encoded_auth_str = base64.encodestring(auth_str)
-            req.add_header("Authorization", "Basic %s" % encoded_auth_str)
-        request = urllib2.urlopen(req)
-        response = request.read()
         
-        metric_count = 0
-        parsed = re.search(r'pong', response)
-        if parsed:
-            metric_count += 1
-            self.gauge("php.ping", 1, tags=tags)
-
-        if metric_count == 0:
-            if self.assumed_url.get(instance['php_ping_url'], None) is None and url[-5:] != '?auto':
-                self.assumed_url[instance['php_ping_url']]= '%s?auto' % url
-                self.warning("Assuming url was not correct. Trying to add ?auto suffix to the url")
-                self.check(instance)
-            else:
-                raise Exception("No metrics were fetched for this instance. Make sure that %s is the proper url." % instance['php_ping_url'])
-
-    @staticmethod
-    def parse_agent_config(agentConfig):
-        if not agentConfig.get('php_ping_url'):
-            return False
-
-        return {
-            'instances': [{'php_ping_url': agentConfig.get('php_ping_url')}]
+        response, content_type = self._get_data(instance)
+        metrics = self.parse_status(response, tags)
+        
+        funcs = {
+            'gauge': self.gauge,
+            'rate': self.rate
         }
+        for row in metrics:
+            try:
+                name, value, tags, metric_type = row
+                func = funcs[metric_type]
+                func(name, value, tags)
+            except Exception:
+                self.log.error(u'Could not submit metric: %s' % repr(row))
+
+    def _get_data(self, instance):
+        url = instance.get('php_ping_url')
+        req = urllib2.Request(url, None, headers(self.agentConfig))
+        if 'php_ping_user' in instance and 'php_ping_password' in instance:
+            add_basic_auth(req, instance['php_ping_user'], instance['php_ping_password'])
+
+        # Submit a service check for status page availability.
+        parsed_url = urlparse.urlparse(url)
+        php_ping_host = parsed_url.hostname
+        php_ping_port = parsed_url.port or 80
+        service_check_name = 'php_ping.can_connect'
+        service_check_tags = ['host:%s' % php_ping_host, 'port:%s' % php_ping_port]
+        try:
+            response = urllib2.urlopen(req)
+        except Exception:
+            self.service_check(service_check_name, AgentCheck.CRITICAL)
+            raise
+        else:
+            self.service_check(service_check_name, AgentCheck.OK)
+
+        body = response.read()
+        resp_headers = response.info()
+        return body, resp_headers.get('Content-Type', 'text/plain')
+
+    @classmethod
+    def parse_status(cls, raw, tags):
+        output = []
+        parsed = re.search(r'pong', raw)
+        if parsed:
+            output.append(('php.ping', 1, tags, 'gauge'))
+
+        return output